refactor(query): Replace tokenRegexp with hand-written scanner
