+++ query.go
@@ -1,9 +1,7 @@
 package tree
 
 import (
-	"encoding/csv"
 	"fmt"
-	"regexp"
 	"strconv"
 	"strings"
 )
@@ -561,277 +559,6 @@ var (
 	_ Selector = (*Comparator)(nil)
 	_ Selector = (*SelectQuery)(nil)
 )
-
-// ParseQuery parses the provided expr to a Query.
-// See https://github.com/mojatter/tree#Query
-func ParseQuery(expr string) (Query, error) {
-	token, err := tokenizeQuery(expr)
-	if err != nil {
-		return nil, err
-	}
-	return tokenToQuery(token, expr)
-}
-
-type token struct {
-	cmd      string
-	method   string
-	argsStr  string
-	quoted   bool
-	value    string
-	parent   *token
-	children []*token
-}
-
-// toValue converts a token to a Node value based on its type and content.
-// Handles quoted strings, booleans, numbers, and defaults to string values.
-func (t *token) toValue() Node {
-	if !t.quoted {
-		if t.value == "" {
-			return Nil
-		}
-		if t.value == "true" {
-			return BoolValue(true)
-		}
-		if t.value == "false" {
-			return BoolValue(false)
-		}
-		if n, err := strconv.ParseFloat(t.value, 64); err == nil {
-			return NumberValue(n)
-		}
-	}
-	return StringValue(t.value)
-}
-
-// indexOfCmd finds the index of a child token with the specified command.
-// Returns -1 if not found.
-func (t *token) indexOfCmd(cmd string) int {
-	for i, c := range t.children {
-		if c.cmd == cmd {
-			return i
-		}
-	}
-	return -1
-}
-
-func (t *token) Args() ([]string, error) {
-	if t.argsStr == "" {
-		return nil, nil
-	}
-	r := csv.NewReader(strings.NewReader(t.argsStr))
-	args, err := r.Read()
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse args: %w", err)
-	}
-	return args, nil
-}
-
-var tokenRegexp = regexp.MustCompile(`"([^"]*)"|(and|or|==|<=|>=|!=|~=|\.\.|[\.\[\]\(\)\|<>:]|([a-z]+)\(([^\)]*)\))|(\w+)`)
-
-// tokenizeQuery parses a query expression string into a token tree.
-// Uses regular expressions to identify different token types.
-func tokenizeQuery(expr string) (*token, error) {
-	current := &token{}
-	ms := tokenRegexp.FindAllStringSubmatch(expr, -1)
-	for _, m := range ms {
-		quoted := m[1]
-		cmd := m[2]
-		method := m[3]
-		argsStr := m[4]
-		word := m[5]
-		// NOTE: detect node name
-		if quoted != "" || word != "" {
-			value := quoted
-			if value == "" {
-				value = word
-			}
-			var lastChild *token
-			if len(current.children) > 0 {
-				lastChild = current.children[len(current.children)-1]
-			}
-			if lastChild != nil && (lastChild.cmd == "." || lastChild.cmd == "..") {
-				lastChild.value = value
-				lastChild.quoted = quoted != ""
-				continue
-			}
-			t := &token{value: value, quoted: quoted != ""}
-			current.children = append(current.children, t)
-			continue
-		}
-		// NOTE: detect keywords
-		t := &token{cmd: cmd, method: method, argsStr: argsStr, parent: current}
-		switch cmd {
-		case "]", ")":
-			if (cmd == "]" && current.cmd != "[") || (cmd == ")" && current.cmd != "(") {
-				return nil, fmt.Errorf("syntax error: no left bracket: %q", expr)
-			}
-			current = current.parent
-		case "[", "(":
-			current.children = append(current.children, t)
-			current = t
-		default:
-			current.children = append(current.children, t)
-		}
-	}
-	if current.parent != nil {
-		return nil, fmt.Errorf("syntax error: no right brackets: %q", expr)
-	}
-	return current, nil
-}
-
-// tokenToQuery converts a token tree into a Query object.
-// Recursively processes tokens based on their command type.
-func tokenToQuery(t *token, expr string) (Query, error) {
-	if t.method != "" {
-		args, err := t.Args()
-		if err != nil {
-			return nil, err
-		}
-		return NewMethodQuery(t.method, args...)
-	}
-	child := len(t.children)
-	switch t.cmd {
-	case "":
-		if child == 0 {
-			return ValueQuery{t.toValue()}, nil
-		}
-	case "|":
-		return SlurpQuery{}, nil
-	case ".":
-		if t.value != "" {
-			return MapQuery(t.value), nil
-		}
-		return NopQuery{}, nil
-	case "..":
-		if t.value != "" {
-			return WalkQuery(t.value), nil
-		}
-		return NopQuery{}, nil
-	case "[":
-		if child == 0 {
-			return SelectQuery{}, nil
-		}
-		if child == 1 {
-			i, err := strconv.Atoi(t.children[0].value)
-			if err != nil {
-				return nil, fmt.Errorf("syntax error: invalid array index: %q", expr)
-			}
-			return ArrayQuery(i), nil
-		}
-		if i := t.indexOfCmd(":"); i != -1 {
-			return tokensToArrayRangeQuery(t.children, i, expr)
-		}
-		selector, err := tokensToSelector(t.children, expr)
-		if err != nil {
-			return nil, err
-		}
-		return SelectQuery{selector}, nil
-	}
-	if child == 0 {
-		return nil, fmt.Errorf("syntax error: invalid token %s: %q", t.cmd, expr)
-	}
-	if child == 1 {
-		return tokenToQuery(t.children[0], expr)
-	}
-	var fq FilterQuery
-	for _, c := range t.children {
-		q, err := tokenToQuery(c, expr)
-		if err != nil {
-			return nil, err
-		}
-		fq = append(fq, q)
-	}
-	return fq, nil
-}
-
-// tokensToArrayRangeQuery creates an ArrayRangeQuery from tokens.
-// Handles array slice notation like [from:to].
-func tokensToArrayRangeQuery(ts []*token, i int, expr string) (Query, error) {
-	from := -1
-	to := -1
-	if j := i - 1; j >= 0 {
-		var err error
-		from, err = strconv.Atoi(ts[j].value)
-		if err != nil {
-			return nil, fmt.Errorf("syntax error: invalid array range: %q", expr)
-		}
-	}
-	if j := i + 1; j < len(ts) {
-		var err error
-		to, err = strconv.Atoi(ts[j].value)
-		if err != nil {
-			return nil, fmt.Errorf("syntax error: invalid array range: %q", expr)
-		}
-	}
-	return ArrayRangeQuery{from, to}, nil
-}
-
-// tokensToSelector converts tokens into a Selector for filtering operations.
-// Handles logical operators (and/or) and comparison operators.
-func tokensToSelector(ts []*token, expr string) (Selector, error) {
-	andOr := ""
-	var groups [][]*token
-	off := 0
-	for i, t := range ts {
-		switch t.cmd {
-		case "and", "or":
-			if andOr != "" && andOr != t.cmd {
-				return nil, fmt.Errorf("syntax error: mixed and|or: %q", expr)
-			}
-			andOr = t.cmd
-			groups = append(groups, ts[off:i])
-			off = i + 1
-		case "(":
-			groups = append(groups, ts[off:i])
-			groups = append(groups, []*token{t})
-			off = i + 1
-		}
-	}
-	groups = append(groups, ts[off:])
-
-	var ss []Selector
-	for _, group := range groups {
-		op := -1
-	GROUP:
-		for i, t := range group {
-			if t.cmd == "(" {
-				sss, err := tokensToSelector(t.children, expr)
-				if err != nil {
-					return nil, err
-				}
-				ss = append(ss, sss)
-				break
-			}
-			switch Operator(t.cmd) {
-			case EQ, GT, GE, LT, LE, NE, RE:
-				op = i
-				break GROUP
-			}
-		}
-		if op == -1 {
-			if len(groups) == 1 && len(group) > 0 {
-				q, err := tokenToQuery(&token{children: group}, expr)
-				if err != nil {
-					return nil, err
-				}
-				ss = append(ss, Evaluator{Query: q})
-			}
-			continue
-		}
-		left, err := tokenToQuery(&token{children: group[0:op]}, expr)
-		if err != nil {
-			return nil, err
-		}
-		right, err := tokenToQuery(&token{children: group[op+1:]}, expr)
-		if err != nil {
-			return nil, err
-		}
-		ss = append(ss, Comparator{left, Operator(group[op].cmd), right})
-	}
-	if andOr == "or" {
-		return Or(ss), nil
-	}
-	return And(ss), nil
-}
 
 // Find finds a node from n using the Query.
 func Find(n Node, expr string) ([]Node, error) {
